requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
langchain>=0.1.0
langchain-huggingface>=0.0.3
langchain-community>=0.0.38
python-dotenv>=1.0.0
pydantic>=2.5.0
# Hugging Face dependencies
transformers>=4.36.0
torch>=2.1.0
sentence-transformers>=2.2.0
accelerate>=0.25.0
# Database
psycopg2-binary>=2.9.5
sqlalchemy>=2.0.0
pgvector>=0.2.0
langchain-postgres>=0.0.5